feat: Pattern builder fluent
correction Pattern builder fluent
