+++ src/main/java/org/api/excel/core/annotations/builder/BoxBuilder.java
@@ -1,0 +1,12 @@
+package org.api.excel.core.annotations.builder;
+
+public class BoxBuilder {
+    private BoxBuilder() {
+
+    }
+
+    public static BoxStep aNew() {
+        return new StepBoxBuilder();
+    }
+
+}
