+++ src/main/java/org/api/excel/mapping/ModelMapper.java
@@ -41,10 +41,10 @@ public class ModelMapper {
         //
         List<CellModel> cellModels = to(fieldContainAnnotation.orElse(null));
         if (bookAnnotationSheets.isPresent()) {
-            return BookModel.aNew().bookAnnotationSheet(bookAnnotationSheets.get()).cellModels(cellModels).create();
+            return BookModel.aNew().book(bookAnnotationSheets.get()).cellModels(cellModels).create();
         }
         return BookModel.aNew()
-                .pageAnnotationSheet(pageAnnotationSheet.orElseThrow())
+                .page(pageAnnotationSheet.orElseThrow())
                 .cellModels(
                         cellModels).create();
     }
