+++ src/main/java/org/api/excel/model/RowHeaderReader.java
@@ -7,14 +7,14 @@ public class RowHeaderReader {
     private final int number;
     private final String columnName;
 
-    private RowHeaderReader(String nameField, int number, String columnName) {
+    RowHeaderReader(String nameField, int number, String columnName) {
         this.nameField = nameField;
         this.number = number;
         this.columnName = columnName;
     }
 
-    public static NameFieldStep aNew() {
-        return new StepRowHeaderReader();
+    public static RowHeaderReaderFluent aNew() {
+        return new RowHeaderReaderBuilder();
     }
 
     public String getNameField() {
@@ -38,49 +38,5 @@ public class RowHeaderReader {
                 .toString();
     }
 
-    public interface NameFieldStep {
-        NumberStep nameField(String nameField);
-    }
-
-    public interface NumberStep {
-        ColumnNameStep number(int number);
-    }
-
-    public interface ColumnNameStep {
-        CreateStep columnName(String columnName);
-    }
-
-    public interface CreateStep {
-        RowHeaderReader create();
-    }
-
-    private static class StepRowHeaderReader implements NameFieldStep, NumberStep, ColumnNameStep, CreateStep {
-        private String nameField;
-        private int number;
-        private String columnName;
-
-        @Override
-        public NumberStep nameField(String nameField) {
-            this.nameField = nameField;
-            return this;
-        }
-
-        @Override
-        public ColumnNameStep number(int number) {
-            this.number = number;
-            return this;
-        }
-
-        @Override
-        public CreateStep columnName(String columnName) {
-            this.columnName = columnName;
-            return this;
-        }
-
-        @Override
-        public RowHeaderReader create() {
-            return new RowHeaderReader(nameField, number, columnName);
-        }
-    }
 
 }
