+++ src/main/java/org/api/excel/model/RowHeaderReaderFluent.java
@@ -1,0 +1,19 @@
+package org.api.excel.model;
+
+public interface RowHeaderReaderFluent {
+
+    Number nameField(String nameField);
+
+
+    interface Number {
+        ColumnName number(int number);
+    }
+
+    interface ColumnName {
+        Create columnName(String columnName);
+    }
+
+    interface Create {
+        RowHeaderReader create();
+    }
+}
