+++ src/main/java/org/api/excel/model/SheetWriterModelFluent.java
@@ -1,0 +1,13 @@
+package org.api.excel.model;
+
+import org.api.excel.core.annotations.Page;
+
+public interface SheetWriterModelFluent {
+    CreateStep pages(Page[] pages);
+
+    interface CreateStep {
+        SheetWriterModel create();
+
+
+    }
+}
