+++ src/main/java/org/api/excel/model/commun/CellModel.java
@@ -3,24 +3,23 @@ package org.api.excel.model.commun;
 import org.api.excel.core.annotations.Box;
 
 import java.lang.reflect.Field;
-import java.util.Objects;
 
 public class CellModel {
     private final Field field;
     private final Box annotation;
 
 
-    private CellModel(Field field, Box annotation) {
+    CellModel(Field field, Box annotation) {
         this.field = field;
         this.annotation = annotation;
     }
 
-    public static StepField aNew() {
-        return new StepCellModel();
+    public static CellModelFluent aNew() {
+        return new CellModelBuilder();
     }
 
-    public static StepDuplicate duplicate(CellModel cellModelOld) {
-        return new StepDuplicateCellModel(cellModelOld.getField(), cellModelOld.getAnnotation());
+    public static CellModelDuplicateFluent duplicate(CellModel cellModelOld) {
+        return new CellModelDuplicate(cellModelOld.getField(), cellModelOld.getAnnotation());
     }
 
     public Field getField() {
@@ -29,64 +28,6 @@ public class CellModel {
 
     public Box getAnnotation() {
         return annotation;
-    }
-
-    public interface StepField {
-        StepBuilderAnnotation field(Field field);
-    }
-
-
-    public interface StepBuilderAnnotation {
-        CellModel create();
-    }
-
-    public interface StepDuplicate {
-        StepDuplicateDone annotation(Box annotation);
-    }
-
-    public interface StepDuplicateDone {
-        CellModel done();
-    }
-
-    private static class StepDuplicateCellModel implements StepDuplicate, StepDuplicateDone {
-        private Field field;
-        private Box annotation;
-
-        public StepDuplicateCellModel(Field field, Box annotation) {
-            this.field = field;
-            this.annotation = annotation;
-        }
-
-        @Override
-        public StepDuplicateCellModel annotation(Box annotation) {
-            this.annotation = annotation;
-            return this;
-        }
-
-        @Override
-        public CellModel done() {
-            Objects.requireNonNull(this.field, "the field cannot is null");
-            Objects.requireNonNull(this.annotation, "the annotation cannot is null");
-            return new CellModel(this.field, this.annotation);
-        }
-    }
-
-    private static class StepCellModel implements StepField, StepBuilderAnnotation {
-        private Field field;
-
-        @Override
-        public StepCellModel field(Field field) {
-            this.field = field;
-            return this;
-        }
-
-        @Override
-        public CellModel create() {
-            Objects.requireNonNull(this.field, "the field cannot is null");
-            Box annotation = field.getAnnotation(Box.class);
-            Objects.requireNonNull(annotation, "the annotation cannot is null");
-            return new CellModel(this.field, annotation);
-        }
     }
 
 }
