+++ src/main/java/org/api/excel/model/commun/CellModelFluent.java
@@ -1,0 +1,11 @@
+package org.api.excel.model.commun;
+
+import java.lang.reflect.Field;
+
+public interface CellModelFluent {
+    Create field(Field field);
+
+    interface Create {
+        CellModel create();
+    }
+}
