+++ src/test/java/org/api/excel/model/PageModelTest.java
@@ -21,7 +21,7 @@ class PageModelTest {
      */
     @Test
     void build_then_notProperties_when_NullPointerException() {
-        Assertions.assertThatThrownBy(BookModel.aNew().pageAnnotationSheet(null).cellModels(Collections.emptyList())::create)
+        Assertions.assertThatThrownBy(BookModel.aNew().page(null).cellModels(Collections.emptyList())::create)
                 .isInstanceOf(NullPointerException.class)
                 .hasMessage("the annotation cannot is null");
 
@@ -36,7 +36,7 @@ class PageModelTest {
      */
     @Test
     void build_then_sheetAnnotationIsNull_when_NullPointerException() {
-        Assertions.assertThatThrownBy(BookModel.aNew().bookAnnotationSheet(null).cellModels(Collections.emptyList())::create)
+        Assertions.assertThatThrownBy(BookModel.aNew().book(null).cellModels(Collections.emptyList())::create)
                 .isInstanceOf(NullPointerException.class)
                 .hasMessage("the annotation cannot is null");
 
@@ -53,7 +53,7 @@ class PageModelTest {
     void build_then_cellModelsIsNull_when_NullPointerException() {
         Optional<Page> classAnnotation = AnnotationInClass.getClassAnnotation(MyClass.class, Page.class);
         Page annotation = classAnnotation.orElse(null);
-        Assertions.assertThatThrownBy(BookModel.aNew().pageAnnotationSheet(annotation).cellModels(null)::create)
+        Assertions.assertThatThrownBy(BookModel.aNew().page(annotation).cellModels(null)::create)
                 .isInstanceOf(NullPointerException.class)
                 .hasMessage("the Collection cannot be null");
 
@@ -63,7 +63,7 @@ class PageModelTest {
      * Methods under test:
      *
      * <ul>
-     *   <li>{@link BookModel#annotationSheets(Book)#builder()}
+     *   <li>{@link BookModel#aNew()#builder()}
      * </ul>
      */
     @Test
@@ -71,7 +71,7 @@ class PageModelTest {
         Optional<Page> classAnnotation = AnnotationInClass.getClassAnnotation(MyClass.class, Page.class);
 
         Page annotation = classAnnotation.orElse(null);
-        Assertions.assertThatThrownBy(BookModel.aNew().pageAnnotationSheet(annotation).cellModels(Collections.emptyList())::create)
+        Assertions.assertThatThrownBy(BookModel.aNew().page(annotation).cellModels(Collections.emptyList())::create)
                 .isInstanceOf(IllegalArgumentException.class)
                 .hasMessage("the Collection cannot be empty");
